تعديل بعض الاشياء يخي اعمل لي بوش اشتي اعرف اش عملت

+++ resources/views/dashboard/admin/includes/ajax/directoratesRigonsStations/directorates.blade.php
@@ -1,39 +1,38 @@
-@section('script')
-<script>
 
+<script>
 
   // ########################## ( Dirctorates SECTION ) ##############################
 
 
      //  Start fetch New Directorate  
-      function fetchNewDirectorate()
+      function fetchDirectorate()
         {
             $.ajax({
                     type: 'get',
                     url: "{{route('directorate.fetchNewDirectorate')}}",
                     dataType:"json",
                     success: function (data) {   
-                     // console.log(data);
+                      console.log(data);
                     $('#fetchAllDirectorates').html("");
                     $('#select_directorates').html("");
-                      $.each(data.directorates, function (key , directorate) {
-                        $('#fetchAllDirectorates').append('<tr  class="offerRow'+directorate.id+' bg-gray-100 hover:scale-95 transform transition-all ease-in">\
-                                  <td class="p-3">'+directorate.id+'</td>\
-                                  <td class="p-3 text-center">'+directorate.directorate_name+'</td>\
-                                  <td class="p-3 flex justify-evenly ">\
-                                    <a href="#"  directorateId="'+directorate.id+'" class="directorateDelete text-gray-400  hover:text-red-400  ">\
-                                      <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">\
-                                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />\
-                                      </svg>\
-                                    </a>\
-                                    <a href="#" directorateId="'+directorate.id+'"  class="directorateEdit text-gray-400 hover:text-yellow-100  ">\
-                                      <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">\
-                                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"   d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />\
-                                      </svg>\
-                                    </a>\
-                                  </td>\
-                                </tr>');
-                          $('#select_directorates').append('<option value='+directorate.id+'>'+directorate.directorate_name+'</option>');
+                      $.each(data.directorates, function (key , dir) {
+                        $('#fetchAllDirectorates').append('<tr class="offerRow'+dir.id+' bg-gray-50 hover:scale-95 transform transition-all ease-in">\
+                                                <td class="p-3">'+dir.id+'</td>\
+                                                <td class="p-3 text-center">'+dir.dirName+'</td>\
+                                                <td class="p-3 flex justify-evenly ">\
+                                                    <a href="#" dirId="'+dir.id+'" class="directorateDelete text-red-400  hover:text-red-600  ">\
+                                                        <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">\
+                                                            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"  d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />\
+                                                        </svg>\
+                                                    </a>\
+                                                    <a href="#" dirId="'+dir.id +'" class="directorateEdit text-yellow-400 hover:text-yellow-600  ">\
+                                                        <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">\
+                                                            <path d="M13.586 3.586a2 2 0 112.828 2.828l-.793.793-2.828-2.828.793-.793zM11.379 5.793L3 14.172V17h2.828l8.38-8.379-2.83-2.828z" />\
+                                                        </svg>\
+                                                    </a>\
+                                                </td>\
+                                            </tr>');
+                          $('#select_directorates').append('<option value='+dir.id+'>'+dir.dirName+'</option>');
                       });
                     
                     
@@ -46,7 +45,7 @@
     // Start Add Directorate By Ajax 
         $(document).on('click', '#saveDirectorate', function (e) {
             e.preventDefault();
-            var formData = new FormData($('#directorateForm')[0]);       
+            var formData = new FormData($('#dirForm')[0]);       
             $.ajax({
                 type: 'POST',
                 enctype: 'multipart/form-data',
@@ -56,11 +55,12 @@
                 contentType: false,
                 cache: false,
                 success: function (data) {
+                  console.log(data);
                    if (data.status == true) {
                      fetchNewDirectorate();
                       alert(data.msgSuccess,'success');
-                        $('#directorateId').val('');
-                        $('#directorateName').val('');
+                        $('#dirId').val('');
+                        $('#dirName').val('');
                         
                      }
                     else alert(data.msgError,'error');
@@ -80,23 +80,26 @@
    
         $(document).on('click', '.directorateDelete', function (e) {
             e.preventDefault();
-              var directorateId = $(this).attr('directorateId');
-            $.ajax({
-                type: 'delete',
-                url:"{{route('directorate.destroy')}}",             
-                data: {
-                     'directorateId' :directorateId, 
-                },
-                success: function (data) {
-                     if (data.status == true) {
-                      alert(data.msgSuccess,'success');
-                      $('.offerRow'+data.directorateId).remove();
-                    }
-                    else alert(data.msgError,'success');
-                }, error: function (reject) {
-                  
-                }
-            });
+            var dirId = $(this).attr('dirId');
+            if(confirm('هل تريد الحدف'))
+            {        
+                    $.ajax({
+                        type: 'delete',
+                        url:"{{route('directorate.destroy')}}",             
+                        data: {
+                            'dirId' :dirId, 
+                        },
+                        success: function (data) {
+                            if (data.status == true) {
+                            alert(data.msgSuccess,'success');
+                            $('.offerRow'+data.dirId).remove();
+                            }
+                            else alert(data.msgError,'success');
+                        }, error: function (reject) {
+                        
+                        }
+                    }); 
+            }
         });
 
     // End Deleting directorate By Ajax
@@ -106,18 +109,18 @@
   
         $(document).on('click', '.directorateEdit', function (e) {
             e.preventDefault();
-              var directorateId = $(this).attr('directorateId');
+              var dirId = $(this).attr('dirId');
             $.ajax({
                 type: 'post',
                 url:"{{route('directorate.edit')}}",
                 data: {
-                     'directorateId' :directorateId, 
+                     'dirId' :dirId, 
                 },
                 success: function (data) {
                   
                      if (data.status == true) {
-                        $('#directorateId').val(data.directorate.id);
-                        $('#directorateName').val(data.directorate.directorate_name);
+                        $('#dirId').val(data.directorate.id);
+                        $('#dirName').val(data.directorate.dirName);
                         window.saveDirectorate.style.display="none";
                         window.updateDirectorate.style.display="inline-flex";
                     }
@@ -137,7 +140,7 @@
 
         $(document).on('click', '#updateDirectorate', function (e) {
             e.preventDefault();
-              var formData = new FormData($('#directorateForm')[0]); 
+              var formData = new FormData($('#dirForm')[0]); 
               
             $.ajax({
                 type: 'post',
@@ -151,8 +154,8 @@
 
                     if(data.status == true){
                         alert(data.msgSuccess,'success');
-                        $('#directorateId').val('');
-                        $('#directorateName').val('');
+                        $('#dirId').val('');
+                        $('#dirName').val('');
                         window.saveDirectorate.style.display="inline-flex";
                         window.updateDirectorate.style.display="none";
                         fetchNewDirectorate();
@@ -172,5 +175,9 @@
 
     //  End Update directorate By Ajax 
 
+
+    
+
 </script>
-@stop
+
+
